+++ CHANGELOG.zh_CN.md
@@ -3,7 +3,8 @@
 ### ✨ Features
 
 - 移除左侧菜单搜索，新增顶部菜单搜索功能
-- layout 移动端适配。页面未适配
+- layout 移动端适配。业务页面未适配
+- axios 加入 joinTime 配置。控制响应是否加入时间戳
 
 ### ⚡ Performance Improvements
 
@@ -19,6 +20,10 @@
 - 刷新按钮布局调整
 - `route.meta` 移除 `externalLink` 属性
 
+### ✨ Refactor
+
+- `openModal`与`openDrawer`第三个参数`openOnSet`默认设置为 true
+
 ### 🐛 Bug Fixes
 
 - 修复多级路由缓存导致组件渲染多次的问题
@@ -31,6 +36,7 @@
 - 修复 `Modal`与`Drawer`组件在使用 emits 数据传递失效问题
 - 修复菜单已知问题
 - 修复上传组件 api 失效问题
+- 修复菜单权限过滤失效问题
 
 ## 2.0.0-rc.13 (2020-12-10)
